added camera and made x, y for instance floats instead of ints

+++ src/engine/Instance.cpp
@@ -1,7 +1,7 @@
 #include "Instance.h"
 
 
-Instance::Instance(int x, int y) {
+Instance::Instance(float x, float y) {
     this->x = x;
     this->y = y;
     this->sprite = new Sprite();
