+++ src/engine/TestObj.cpp
@@ -2,7 +2,7 @@
 #include "TestObj.h"
 
 
-TestObj::TestObj(int x, int y) : Instance(x, y) {
+TestObj::TestObj(float x, float y) : Instance(x, y) {
     this->sprite->addImage(game.loader->load("assets/card.png"));
 }
 
@@ -23,16 +23,16 @@ void TestObj::render(float delta) {
     glTranslatef(this->x, this->y, 0.0f);
     glBegin(GL_QUADS);
     glTexCoord2f(0, 0);
-    glVertex2f( 0.0f, 0.0f );
+    glVertex2f(0.0f, 0.0f);
 
     glTexCoord2f(1, 0);
-    glVertex2f( 0.0f, 200.0f );
+    glVertex2f(0.0f, 200.0f);
 
     glTexCoord2f(1, 1);
-    glVertex2f( 200.0f, 200.0f );
+    glVertex2f(200.0f, 200.0f);
 
     glTexCoord2f(0, 1);
-    glVertex2f( 200.0f, 0.0f );
+    glVertex2f(200.0f, 0.0f);
     glEnd();
     glPopMatrix();
 }
